% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{loadHadleyverse}
\alias{loadHadleyverse}
\title{Loads the Hadleyverse}
\usage{
loadHadleyverse()
}
\description{
This is roughly equivalent to what gets run when this package is loaded, so you won't generally need this function.
}
\examples{
loadHadleyverse()
}

